Simplify installer by using '-o' parameter
When the '-o'  parameter is used together with '-i', the installer will
look for the latest rdpwrap.ini file online, which makes runnig the
installer again with '-w' unnecessary.
If there's no network acces, the installer will fall back to the
built-in one, even if online installation is enabled via '-o'.

This fixes "error 5" while installing on newer versions of Windows.